Move User to From Entites to Entities.Identity
