+++ Capri.Database/Entities/DeanEmployee.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Capri.Database.Entities.Identity;
 
 namespace Capri.Database.Entities
 {
