+++ Capri.Database/Entities/Identity/User.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Identity;
 
-namespace Capri.Database.Entities
+namespace Capri.Database.Entities.Identity
 {
     public class User : IdentityUser<Guid>, IEntity
     {
