+++ Capri.Database/Entities/Student.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using System.ComponentModel.DataAnnotations;
+using Capri.Database.Entities.Identity;
 
 namespace Capri.Database.Entities
 {
