+++ Capri.Web.Services/AccountService.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.Extensions.Options;
 using Capri.Database;
-using Capri.Database.Entities;
+using Capri.Database.Entities.Identity;
 using Capri.Web.Services.Settings;
 using Capri.Web.ViewModels.User;
